requirements/requirements.txt
aioredis==1.2.0
requests==2.21.0
pyrogram[fast]==0.11.0
Pillow==5.4.1
pytz==2018.9
